fastapi/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["bobs24 <bobsebastian1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
fastapi = "^0.110.0"
joblib = "^1.3.2"
scikit-learn = "1.1.3"
asyncpg = "^0.29.0"
tensorflow-io-gcs-filesystem = {version = "*", python = ">=3.11,<3.12"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
